Cambios a calculos de facturacion

+++ ejb/src/main/java/ec/mileniumtech/educafacil/modelo/persistencia/entity/DetalleFactura.java
@@ -39,5 +39,5 @@ public class DetalleFactura {
     private BigDecimal descuento;
     
     @Transient
-    private Integer impuestoIva;
+    private BigDecimal impuestoIva;
 }
